requirements.txt
mysql-connector-python
PyJWT
flask-cors
pandas
scikit-learn
joblib
numpy
xgboost
openai